+++ playground/express/index.ts
@@ -5,13 +5,27 @@ import {
   Controller,
   Response,
   Get,
-  attachControllers
+  attachControllers,
+  ERROR_MIDDLEWARE,
+  ErrorMiddleware,
+  Middleware
 } from '@decorators/express';
 
 const MESSAGE = new InjectionToken('MESSAGE');
 
 @Injectable()
-class Middleware {
+class ServerErrorMiddleware implements ErrorMiddleware {
+  constructor(@Inject(MESSAGE) private message: string) {}
+
+  public use(error, req, res, next) {
+    console.log('server error middleware', this.message);
+
+    res.send(500);
+  }
+}
+
+@Injectable()
+class UserMiddleware implements Middleware {
   constructor(@Inject(MESSAGE) private message: string) {}
 
   public use(req, res, next) {
@@ -27,9 +41,10 @@ class UserController {
 
   constructor(@Inject(MESSAGE) private message: string) {}
 
-  @Get('/user', [Middleware])
+  @Get('/user', [UserMiddleware])
   public getData(@Response() res): void {
-    res.send(this.message);
+    throw Error('User test error');
+    // res.send(this.message);
   }
 
 }
@@ -37,6 +52,7 @@ class UserController {
 let app: express.Express = express();
 
 Container.provide([
+  { provide: ERROR_MIDDLEWARE, useClass: ServerErrorMiddleware },
   { provide: MESSAGE, useValue: 'Express welcomes user' }
 ]);